+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 * Add `SystemdUnitsOptsBuilder::additional_env_variables`
 * Add `SystemdUnitsOptsBuilder::after`
 * Add `SystemdUnitsOptsBuilder::requires`
+* Add `SystemdUnitsOptsBuilder::wants`
 
 # 0.7.0
 * Make futures returned by methods Send + Sync